+++ packages/ensjs/src/functions/subgraph/getNameHistory.ts
@@ -65,6 +65,9 @@ export type GetNameHistoryReturnType = {
 
 /**
  * Gets the history of a name from the subgraph.
+ * It's recommended to use a custom subgraph URL to avoid rate limiting.
+ * @see {@link ../../../../../docs/basics/custom-subgraph-uris.md | Custom Subgraph URIs Guide}
+ *
  * @param client - {@link ClientWithEns}
  * @param parameters - {@link GetNameHistoryParameters}
  * @returns History object, or null if name could not be found. {@link GetNameHistoryReturnType}