+++ packages/ensjs/vitest.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig } from 'vitest/config'
+// TODO: Include subgraph functions before final merge.
 
 export default defineConfig({
   test: {