+++ packages/ensjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ensdomains/ensjs
 
+## 4.2.2
+
+### Patch Changes
+
+- [#301](https://github.com/ensdomains/ensjs/pull/301) [`6458e3d`](https://github.com/ensdomains/ensjs/commit/6458e3d869dc1b84f4acd9835d39ea97a3f6bc62) Thanks [@TateB](https://github.com/TateB)! - Fixed custom error handling in getRecords
+
 ## 4.0.2
 
 ### Patch Changes